assistant-chat/requirement.txt
numpy
scikit-learn
langchain
langchain-community
langchain-core
langgraph
datasets
langchain_groq
transformers
sentence-transformers
faiss-cpu
flask
flask-cors
gunicorn